requirements.txt
asgiref==3.5.2
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
numpy==1.23.5
openpyxl==3.0.10
pandas==1.5.2
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
